requirements.txt
bedrock-agentcore
strands-agents
boto3
botocore
requests
